Реализована часть с сохранением времени, потраченным на задачу

+++ Company.Backend/Core/Company.Application/Interfaces/ITimeService.cs
@@ -4,7 +4,7 @@ namespace Company.Application.Interfaces
 {
     public interface ITimeService
     {
-        int ConvertTimeInMinutes(TimeProjectVm time);
+        long ConvertTimeInMinutes(TimeProjectVm time);
 
         TimeProjectVm ConvertMinutesInTime(int minutes);
     }
